# openapi-spec/openapi-unified.yaml
openapi: 3.0.0
info:
  title: E-Commerce Microservices API
  description: |
    Comprehensive API documentation for the E-Commerce Microservices platform.
    This unified specification documents all microservices APIs including:
    - Customer Service
    - Order Service
    - Inventory Service
    - Payment Service
    - Authentication Service
  version: '1.0.0'
  contact:
    name: E-Commerce API Support
    email: api-support@ecommerce.local
    url: https://ecommerce.local/support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.staging.ecommerce.local
    description: Staging Environment
  - url: https://api.ecommerce.local
    description: Production Environment
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          enum:
            - RESOURCE_NOT_FOUND
            - VALIDATION_FAILED
            - INSUFFICIENT_STOCK
            - EVENT_PUBLISHING_ERROR
            - BUSINESS_ERROR
            - INTERNAL_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - AUTHENTICATION_FAILED
        message:
          type: string
          example: Resource not found
        timestamp:
          type: string
          format: date-time
        details:
          type: object
    PagedResponse:
      type: object
      required:
        - content
        - pageNumber
        - pageSize
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
          example: 0
        pageSize:
          type: integer
          format: int32
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          format: int32
          example: 5
...
    CustomerResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: customer@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phoneNumber:
          type: string
          example: '+1-555-0123'
        address:
          type: string
          example: '123 Main St, Anytown, USA'
        city:
          type: string
          example: Anytown
        country:
          type: string
          example: USA
        postalCode:
          type: string
          example: '12345'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCustomerRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: John
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Doe
        phoneNumber:
          type: string
          example: '+1-555-0123'
        address:
          type: string
          example: '123 Main St, Anytown, USA'
        city:
          type: string
          example: Anytown
        country:
          type: string
          example: USA
        postalCode:
          type: string
          example: '12345'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        RS256 access token issued by Keycloak (realm `ecommerce`). There is no login endpoint on this
        API: obtain a token from Keycloak's token endpoint, then send it as `Authorization: Bearer <token>`.
        See docs/KEYCLOAK_IDENTITY.md for roles and object-level rules.
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        default: id
      example: id
security:
  - BearerAuth: []
paths:
  /api/customers:
    get:
      tags:
        - Customer Management
      summary: Get all customers
      description: Retrieve all customers with pagination support
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        '200':
          description: Successfully retrieved customers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/CustomerResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Customer Management
      summary: Create a new customer
      description: Create a new customer with provided information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/customers/{id}:
    get:
      tags:
        - Customer Management
      summary: Get customer by ID
      description: Retrieve a specific customer by their ID
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Customer Management
      summary: Update customer
      description: Update an existing customer's information
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Customer Management
      summary: Delete customer
      description: Delete a customer by their ID
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...